feat: add eslint configuration file
Add a new file `.eslintrc.json` to the project with the following configuration:
- Set the environment to `browser` and `es6`
- Extend the `eslint:recommended`, `plugin:import/errors`, `plugin:import/warnings`, and `prettier` configurations
- Set the parser options to use ECMAScript 2021 and source type module
- Add the `import` plugin and configure the following rules:
  - `no-unused-vars` to be an error
  - `import/first` to be an error
  - `import/no-duplicates` to be an error
  - `import/newline-after-import` to be an error
  - `import/no-unresolved` to be an error
  - `import/extensions` to be an error with the options `{ "js": "never", "jsx": "never" }`
  - `import/named` to be an error
  - `import/default` to be an error
  - `import/no-absolute-path` to be an error
  - `import/no-webpack-loader-syntax` to be an error

fix: change port variable case from lowercase `port` to uppercase `PORT` to improve semantics

Change the variable `port` to `PORT` in the `server.ts` file to improve semantics and make it consistent with the naming convention.

feat: add support for `process.env.PORT` environment variable to be able to run app on a configurable port

Add support for the `process.env.PORT` environment variable in the `server.ts` file to allow running the app on a configurable port. This will make it easier to deploy the app to different environments without modifying the code.

feat: add `createEmptyBoard` function to `board.js`

Add a new function `createEmptyBoard` to the `board.js` file. This function creates an empty board with the specified width and height. It is used to initialize the game board.

feat: add event listeners for keyboard controls in `control.js`

Add event listeners for keyboard controls in the `control.js` file. The controls include arrow keys for movement, space key for rotation, and arrow up key for dropping the piece to the bottom.

feat: add `draw` and `drawNextPiece` functions to `draw.js`

Add the `draw` and `drawNextPiece` functions to the `draw.js`